trying to get intermediate values as floats

+++ Labs/lab2/mykernel.cl
@@ -15,8 +15,8 @@ __kernel void calculate_pi(
    /* Perform work item calculation */
    for(int i = 1; i < num_terms*2; i+=4){
       
-      float term_1 = 1/((work_item_index*8)+i);
-      float term_2 = 1/((work_item_index*8)+i+2);
+      float term_1 = (float)1/((work_item_index*8.0)+i);
+      float term_2 = (float)1/((work_item_index*8.0)+i+2);
       printf("Work Item #%d: term 1:%f, term2:%f",work_item_index,term_1,term_2);
       
       local_result[work_item_index] += (term_1 - term_2);
